+++ CHANGELOG.md
@@ -17,6 +17,8 @@ All notable changes to this project will be documented in this file.
 - `plot_options` not deeply merged on bar and column chart when
   `options` contains `plot_options`
 - Undefined method `to_json` for `options` when not using Rails
+- Utils::Hash.camelize_keys not camelizing keys of hash in array
+  value
 
 
 ## [0.1.4] - 2019-07-06